services/web/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asarPy==1.0.1
asn1crypto==1.5.1
blinker==1.9.0
cbor2==5.6.5
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security==5.5.2
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.0
gunicorn==23.0.0
MarkupSafe
SQLAlchemy
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
numpy
opencv-python==4.10.0.84
packaging==24.2
passlib==1.7.4
pycparser==2.22
pyOpenSSL==24.3.0
python-dotenv==1.0.1
qrcode==8.0
sensor==6
typing_extensions==4.12.2
webauthn==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
